Restructure views and models
