+++ app/docs/userregister.yml
@@ -1,0 +1,53 @@
+Register a user
+This resource registers new users
+---
+tags:
+  - Authentication
+parameters:
+  - name: body
+    in: body
+    type: string
+    required: true
+    schema:
+        id: registration_input
+        properties:
+            username:
+                type: string
+                description: user's nick name
+            name:
+                type: string
+                description: user's full name, firstname and last name
+            password:
+                type: string
+                description: user' password for authentication
+responses:
+  201:
+    description: A user's account created
+    content: application/json
+    schema:
+        id: registration_output
+        properties:
+            username:
+                type: string
+                description: The username
+                default: some_username
+  400:
+    description: A bad request mainly as a result of invalid data
+    content: application/json
+    schema:
+        id: registration_output_400
+        properties:
+            Error:
+                type: string
+                description: Error
+                default: some_error
+  409:
+    description: Conflict, a similar record already exists
+    content: application/json
+    schema:
+        id: registration_output_409
+        properties:
+            Error:
+                type: string
+                description: Error
+                default: The username already exists
