+++ app/docs/userresetpassword.yml
@@ -1,0 +1,64 @@
+
+Reset password
+This resource resets a registered user's password
+---
+tags:
+  - Authentication
+parameters:
+  - name: body
+    in: body
+    type: string
+    required: true
+    schema:
+        id: reset_password_input
+        properties:
+            password:
+                type: string
+                description: user's authentication password
+            new_password:
+                type: string
+                description: user's new authentication password
+            confirm_password:
+                type: string
+                description: user's matching new authentication password
+responses:
+  201:
+    description: Login successful
+    content: application/json
+    schema:
+      id: reset_password_output
+      properties:
+        message:
+          type: string
+          description: Success message
+          default: Your password was successfully reset
+  400:
+    description: A bad request mainly as a result of invalid data
+    content: application/json
+    schema:
+      id: reset_password_output_400
+      properties:
+        Error:
+          type: string
+          description: Error
+          default: some_error
+  401:
+    description: Un authorized, the resource requires authorization
+    content: application/json
+    schema:
+      id: reset_password_output_401
+      properties:
+        Error:
+          type: string
+          description: Error
+          default: some_error
+  403:
+    description: Invalid request mainly requiering permissions
+    content: application/json
+    schema:
+      id: reset_password_output_403
+      properties:
+        Error:
+          type: string
+          description: Error
+          default: Incorrect password
