+++ tests/test_delete_account.py
@@ -5,6 +5,7 @@ import json
 class TestDeleteAccount(BaseTestCase):
     # ---delete account tests
     def test_deleting_account_without_password_key(self):
+        """Test deletinga user account without a password key"""
         self.create_user()
         self.create_category()
         self.create_recipe()
@@ -18,6 +19,7 @@ class TestDeleteAccount(BaseTestCase):
                              'password key and value')
 
     def test_deleting_account_with_empty_password_value(self):
+        """Test deleting an account with emoty password value"""
         self.create_user()
         with self.client:
             headers = self.helper_login_with_token()
@@ -30,6 +32,7 @@ class TestDeleteAccount(BaseTestCase):
             self.assertEqual(reply['Error'], 'password is empty')
 
     def test_deleting_account_with_wrong_password(self):
+        """Test deleting an account with wrong password"""
         self.create_user()
         with self.client:
             headers = self.helper_login_with_token()
@@ -42,6 +45,7 @@ class TestDeleteAccount(BaseTestCase):
             self.assertEqual(reply['Error'], 'Incorrect password')
 
     def test_deleting_account_successfully(self):
+        """Test deleting an account successfully"""
         self.create_user()
         with self.client:
             headers = self.helper_login_with_token()
@@ -49,5 +53,6 @@ class TestDeleteAccount(BaseTestCase):
                                           content_type='application/json',
                                           headers=headers,
                                           data=json.dumps(
-                                              dict(password='pass')))
+                                              dict(password=self.test_user_password
+                                                   )))
             self.assertEqual(response.status_code, 204)
